pyproject.toml
[project]
name = "nextlabs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django-cors-headers>=4.6.0",
    "django>=5.1.5",
    "djangorestframework>=3.15.2",
    "ruff>=0.9.3",
    "djangorestframework-simplejwt>=5.4.0",
    "pillow>=11.1.0",
]
